Download csv for each season
I encourage you to not actually scrape the full 20+seasons as shown in the scrape file as this burdens basketball.ref.com. Instead, just run these calls in your r environment and you should be good. The scrape walk through is more for learning and adapting to other types of data you might want to scrape in bball ref or other sports ref sites